feat: add database configs in persistence.xml and pom.xml files & fix imports of Product
