--- src/main/java/uol/compass/commerce/Application.java
@@ -1,37 +1,0 @@
-package uol.compass.commerce;
-
-import java.sql.SQLException;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
-
-import org.h2.tools.Server;
-
-public class Application {
-
-    public static void main(String[] args) throws SQLException {
-        startDatabase();
-
-        final EntityManagerFactory entityManagerFactory = Persistence.createEntityManagerFactory("uol.compass.commerce");
-        final EntityManager entityManager = entityManagerFactory.createEntityManager();
-
-        Product product = new Product();
-        product.setProductName("Copo Stanley");
-        product.setProductDescription("Copo para tomar cerveja");
-
-        entityManager.getTransaction().begin();
-        entityManager.persist(product);
-        entityManager.getTransaction().commit();
-
-        System.out.println("Open your browser and navigate to http://localhost:8082/");
-        System.out.println("JDBC URL: jdbc:h2:mem:my_database");
-        System.out.println("User Name: sa");
-        System.out.println("Password: ");
-
-    }
-
-    private static void startDatabase() throws SQLException {
-        new Server().runTool("-tcp", "-web", "-ifNotExists");
-    }
-}
